event/cnxk: fix getwork mode devargs parsing
[ upstream commit 343216ff0c9f89d8d31060c53ad7271c87bc1323 ]

CN10K can support up to three types of getwork modes, use parse value
function to parse the device arguments.

Fixes: 5fb651552c79 ("event/cnxk: fix SSO and TIM argument parsing")

+++ drivers/event/cnxk/cnxk_eventdev.h
@@ -113,7 +113,7 @@ struct cnxk_sso_evdev {
 	/* CN9K */
 	uint8_t dual_ws;
 	/* CN10K */
-	uint8_t gw_mode;
+	uint32_t gw_mode;
 	/* Crypto adapter */
 	uint8_t is_ca_internal_port;
 } __rte_cache_aligned;
